refactor(examples/web-email): remove unused
